Move Background + BackgroundCache defs to ui module
